requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1 